# Flow executor speedups and tests