--- a/flow/analyze.go
+++ b/flow/analyze.go
@@ -12,9 +12,9 @@ func OrderEdgesByEdgeAttributeOrderOrNodeKey(a, b xg.Edge) bool {
 	ca := a.Attributes()
 	cb := b.Attributes()
 	if len(ca) > 0 && len(cb) > 0 {
-		idx, ok := ca["order"].(int)
+		idx, ok := ca["arg"].(int)
 		if ok {
-			idx2, ok2 := cb["order"].(int)
+			idx2, ok2 := cb["arg"].(int)
 			if ok2 {
 				return idx < idx2
 			}
@@ -25,6 +25,10 @@ func OrderEdgesByEdgeAttributeOrderOrNodeKey(a, b xg.Edge) bool {
 
 func analyze(ref GraphRef, g xg.Graph, kind xg.EdgeKind, ordered xg.NodeSlice,
 	options Options) (*graph, error) {
+
+	if options.Logger == nil {
+		options.Logger = nologging{}
+	}
 
 	nodes := []*node{}
 	links := map[xg.Edge]chan work{}