--- a/flow/testing.go
+++ b/flow/testing.go
@@ -2,51 +2,9 @@ package flow // import "github.com/orkestr8/xgraph/flow"
 
 import (
 	"fmt"
-	"testing"
 
 	xg "github.com/orkestr8/xgraph"
 )
-
-const (
-	// On slower machines testing.T.Log() fails when `go test -race`.
-	useTestingLog = true
-)
-
-type testlog struct {
-	*testing.T
-}
-
-func (s testlog) Log(context string, args ...interface{}) {
-	if useTestingLog {
-		s.T.Log([]interface{}{"INFO", context, args}...)
-		return
-	}
-	fmt.Println([]interface{}{"INFO", context, args}...)
-}
-
-func (s testlog) Warn(context string, args ...interface{}) {
-	if useTestingLog {
-		s.T.Log([]interface{}{"WARN", context, args}...)
-	}
-	fmt.Println([]interface{}{"WARN", context, args}...)
-}
-
-type benchlog struct {
-	*testing.B
-	log bool
-}
-
-func (s benchlog) Log(context string, args ...interface{}) {
-	if s.log {
-		s.B.Log([]interface{}{"INFO", context, args}...)
-	}
-}
-
-func (s benchlog) Warn(context string, args ...interface{}) {
-	if s.log {
-		s.B.Log([]interface{}{"WARN", context, args}...)
-	}
-}
 
 func testBuildGraph(input xg.EdgeKind) xg.Graph {
 
@@ -72,10 +30,10 @@ func testBuildGraph(input xg.EdgeKind) xg.Graph {
 	g.Associate(x1, input, sumX) // ordering comes from the nodeKey, lexicographically
 	g.Associate(x2, input, sumX)
 	g.Associate(x3, input, sumX)
-	g.Associate(y1, input, sumY, xg.Attribute{Key: "order", Value: 2})
-	g.Associate(y2, input, sumY, xg.Attribute{Key: "order", Value: 1})
-	g.Associate(x3, input, sumY, xg.Attribute{Key: "order", Value: 0})
-	g.Associate(sumX, input, ratio, xg.Attribute{Key: "order", Value: 0}) // positional arg index
-	g.Associate(sumY, input, ratio, xg.Attribute{Key: "order", Value: 1})
+	g.Associate(y1, input, sumY, xg.Attribute{Key: "arg", Value: 2})
+	g.Associate(y2, input, sumY, xg.Attribute{Key: "arg", Value: 1})
+	g.Associate(x3, input, sumY, xg.Attribute{Key: "arg", Value: 0})
+	g.Associate(sumX, input, ratio, xg.Attribute{Key: "arg", Value: 0}) // positional arg index
+	g.Associate(sumY, input, ratio, xg.Attribute{Key: "arg", Value: 1})
 	return g
 }